+++ edgelet/support-bundle/src/runtime_util.rs
@@ -1,10 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 
-use std::env;
 use std::io::Write;
-
-// TODO: make tokio
-use std::process::Command as ShellCommand;
 
 use failure::Fail;
 use futures::StreamExt;
@@ -30,39 +26,6 @@ pub async fn get_modules(
     Ok(runtime_modules)
 }
 
-pub async fn write_check(
-    writer: &mut impl Write,
-    iothub_hostname: Option<String>,
-) -> Result<(), Error> {
-    print_verbose("Calling iotedge check");
-
-    let mut iotedge = env::args().next().unwrap();
-    if iotedge.contains("aziot-edged") {
-        print_verbose("Calling iotedge check from edgelet, using iotedge from path");
-        iotedge = "iotedge".to_string();
-    }
-
-    let mut check = ShellCommand::new(iotedge);
-    check.arg("check").args(&["-o", "json"]);
-
-    if let Some(host_name) = iothub_hostname {
-        check.args(&["--iothub-hostname", &host_name]);
-    }
-    let check = check
-        .output()
-        .map_err(|err| Error::from(err.context(ErrorKind::SupportBundle)))?;
-
-    writer
-        .write_all(&check.stdout)
-        .map_err(|err| Error::from(err.context(ErrorKind::SupportBundle)))?;
-    writer
-        .write_all(&check.stderr)
-        .map_err(|err| Error::from(err.context(ErrorKind::SupportBundle)))?;
-
-    print_verbose("Wrote check output to file");
-    Ok(())
-}
-
 pub async fn write_logs(
     runtime: &impl ModuleRuntime,
     module_name: &str,
@@ -85,13 +48,4 @@ pub async fn write_logs(
         .into_iter()
         .collect::<Result<(), std::io::Error>>()
         .map_err(|err| Error::from(err.context(ErrorKind::Write)))
-}
-
-fn print_verbose<S>(message: S)
-where
-    S: std::fmt::Display,
-{
-    if true {
-        println!("{}", message);
-    }
 }
